# Unifies, fixes, checks and benchmarks `ImageBuffer` index calculation