--- a/src/buffer.rs
+++ b/src/buffer.rs
@@ -831,7 +831,37 @@ mod test {
 
     #[bench]
     #[cfg(feature = "benchmarks")]
-    fn bench_image_access_good_cache(b: &mut test::Bencher) {
+    fn bench_image_access_row_by_row(b: &mut test::Bencher) {
+        use buffer::{ImageBuffer, Pixel};
+
+        let mut a: RgbImage = ImageBuffer::new(1000, 1000);
+        for mut p in a.pixels_mut() {
+            let rgb = p.channels_mut();
+            rgb[0] = 255;
+            rgb[1] = 23;
+            rgb[2] = 42;
+        }
+
+        b.iter(move || {
+            let image: &RgbImage = test::black_box(&a);
+            let mut sum: usize = 0;
+            for y in 0..1000 {
+                for x in 0..1000 {
+                    let pixel = image.get_pixel(x, y);
+                    sum = sum.wrapping_add(pixel[0] as usize);
+                    sum = sum.wrapping_add(pixel[1] as usize);
+                    sum = sum.wrapping_add(pixel[2] as usize);
+                }
+            }
+            test::black_box(sum)
+        });
+
+        b.bytes = 1000 * 1000 * 3;
+    }
+
+    #[bench]
+    #[cfg(feature = "benchmarks")]
+    fn bench_image_access_col_by_col(b: &mut test::Bencher) {
         use buffer::{ImageBuffer, Pixel};
 
         let mut a: RgbImage = ImageBuffer::new(1000, 1000);