Fixes #836 (and by extension addresses one of the issues in https://github.com/PistonDevelopers/imageproc/issues/281). I'll leave this open for a few more days or hours in case an additional idea for a performance improvement comes up (thanks @fintelia for the current calculation). Right now:

```
Before: test … bench:   1,541,207 ns/iter (+/- 224,258) = 1946 MB/s
After: test … bench:   1,759,194 ns/iter (+/- 62,646) = 1705 MB/s
```

That seems acceptable but maybe not yet optimal.