@HeroicKatora 2019-01-15
Ehm, this fails some tests? Note that tests are only executed for nightly on Travis.